+++ qiskit/test/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2018, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
+"""Facilities for testing Qiskit."""
+
+from .common import QiskitTestCase
+from .decorators import requires_cpp_simulator, requires_qe_access, slow_test
+from .utils import Path
