+++ qiskit/test/decorators.py
@@ -1,0 +1,164 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2018, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
+"""Decorator for using with Qiskit unit tests."""
+
+import functools
+import os
+import unittest
+
+from qiskit.providers.ibmq.credentials import Credentials, discover_credentials
+from qiskit.providers.legacysimulators import QasmSimulator
+
+from .utils import Path
+from .http_recorder import http_recorder
+from .testing_options import get_test_options
+
+
+def is_cpp_simulator_available():
+    """Check if the C++ simulator can be instantiated.
+
+    Returns:
+        bool: True if simulator executable is available
+    """
+    try:
+        QasmSimulator()
+    except FileNotFoundError:
+        return False
+    return True
+
+
+def requires_cpp_simulator(test_item):
+    """Decorator that skips test if C++ simulator is not available
+
+    Args:
+        test_item (callable): function or class to be decorated.
+
+    Returns:
+        callable: the decorated function.
+    """
+    reason = 'C++ simulator not found, skipping test'
+    return unittest.skipIf(not is_cpp_simulator_available(), reason)(test_item)
+
+
+def slow_test(func):
+    """
+    Decorator that signals that the test takes minutes to run.
+
+    Args:
+        func (callable): test function to be decorated.
+
+    Returns:
+        callable: the decorated function.
+    """
+
+    @functools.wraps(func)
+    def _wrapper(*args, **kwargs):
+        skip_slow = not TEST_OPTIONS['run_slow']
+        if skip_slow:
+            raise unittest.SkipTest('Skipping slow tests')
+
+        return func(*args, **kwargs)
+
+    return _wrapper
+
+
+def _get_credentials(test_object, test_options):
+    """
+    Finds the credentials for a specific test and options.
+
+    Args:
+        test_object (QiskitTestCase): The test object asking for credentials
+        test_options (dict): Options after QISKIT_TESTS was parsed by get_test_options.
+
+    Returns:
+        Credentials: set of credentials
+
+    Raises:
+        Exception: When the credential could not be set and they are needed for that set of options
+    """
+
+    dummy_credentials = Credentials('dummyapiusersloginWithTokenid01',
+                                    'https://quantumexperience.ng.bluemix.net/api')
+
+    if test_options['mock_online']:
+        return dummy_credentials
+
+    if os.getenv('USE_ALTERNATE_ENV_CREDENTIALS', ''):
+        # Special case: instead of using the standard credentials mechanism,
+        # load them from different environment variables. This assumes they
+        # will always be in place, as is used by the Travis setup.
+        return Credentials(os.getenv('IBMQ_TOKEN'), os.getenv('IBMQ_URL'))
+    else:
+        # Attempt to read the standard credentials.
+        discovered_credentials = discover_credentials()
+
+        if discovered_credentials:
+            # Decide which credentials to use for testing.
+            if len(discovered_credentials) > 1:
+                try:
+                    # Attempt to use QE credentials.
+                    return discovered_credentials[dummy_credentials.unique_id()]
+                except KeyError:
+                    pass
+
+            # Use the first available credentials.
+            return list(discovered_credentials.values())[0]
+
+    # No user credentials were found.
+    if test_options['rec']:
+        raise Exception('Could not locate valid credentials. You need them for recording '
+                        'tests against the remote API.')
+
+    test_object.log.warning("No user credentials were detected. Running with mocked data.")
+    test_options['mock_online'] = True
+    return dummy_credentials
+
+
+def requires_qe_access(func):
+    """
+    Decorator that signals that the test uses the online API:
+        * determines if the test should be skipped by checking environment
+            variables.
+        * if the `USE_ALTERNATE_ENV_CREDENTIALS` environment variable is
+          set, it reads the credentials from an alternative set of environment
+          variables.
+        * if the test is not skipped, it reads `qe_token` and `qe_url` from
+            `Qconfig.py`, environment variables or qiskitrc.
+        * if the test is not skipped, it appends `qe_token` and `qe_url` as
+            arguments to the test function.
+    Args:
+        func (callable): test function to be decorated.
+
+    Returns:
+        callable: the decorated function.
+    """
+
+    @functools.wraps(func)
+    def _wrapper(self, *args, **kwargs):
+        if TEST_OPTIONS['skip_online']:
+            raise unittest.SkipTest('Skipping online tests')
+
+        credentials = _get_credentials(self, TEST_OPTIONS)
+        self.using_ibmq_credentials = credentials.is_ibmq()
+        kwargs.update({'qe_token': credentials.token,
+                       'qe_url': credentials.url})
+
+        decorated_func = func
+        if TEST_OPTIONS['rec'] or TEST_OPTIONS['mock_online']:
+            # For recording or for replaying existing cassettes, the test
+            # should be decorated with @use_cassette.
+            vcr_mode = 'new_episodes' if TEST_OPTIONS['rec'] else 'none'
+            decorated_func = http_recorder(
+                vcr_mode, Path.CASSETTES.value).use_cassette()(decorated_func)
+
+        return decorated_func(self, *args, **kwargs)
+
+    return _wrapper
+
+
+TEST_OPTIONS = get_test_options()
